Changed styleguidist config file and removed style for html dom element from the component styles.
There was a problem with styleguidist styles, as it didn't inherit font-size from the app global styles, where is reset of the rem unit to be multiple of 10px, not 16px. Problem has been removed by adding custom config. It allowed to remove global styles from the component scss file.